Move Swagger extensions to own class
Move extension method to add Swagger to its own class.
